+++ projects/ngx-material/src/lib/buttons/index.ts
@@ -1,0 +1,3 @@
+
+export * from './components/button/button.component';
+export * from './components/icon-button/icon-button.component';